Move to separate editor app
