+++ control/api_views.py
@@ -15,8 +15,6 @@ from control.permissions import ChangeControlPermission, ChangeQuestionnairePerm
 from .serializers import QuestionSerializer, QuestionnaireSerializer, QuestionnaireUpdateSerializer
 from .serializers import ThemeSerializer, QuestionFileSerializer, ResponseFileSerializer, ResponseFileTrashSerializer
 
-from .serializers import UpdateEditorQuestionnaireSerializer # todo move to separate app
-
 # This signal is triggered after the questionnaire is saved via the API
 questionnaire_api_post_save = django.dispatch.Signal(providing_args=["instance"])
 
@@ -333,15 +331,3 @@ class QuestionnaireViewSet(viewsets.ModelViewSet):
             'target': control,
         }
         action.send(**action_details)
-
-class UpdateEditor(generics.UpdateAPIView):  # todo move to separate app
-    serializer_class = UpdateEditorQuestionnaireSerializer
-
-    def get_queryset(self):
-        if not self.request.user.profile.is_inspector:
-            return Questionnaire.objects.none()
-
-        queryset = Questionnaire.objects  \
-            .filter(control__in=self.request.user.profile.controls.all())  \
-            .filter(is_draft=True)
-        return queryset
